# apps/OpenVulnScan/internal/router/openapi.yaml
openapi: 3.0.3
info:
  title: OpenVulnScan API
  description: |
    ## OpenVulnScan — Vulnerability Scanning & Management Platform

    Built as a Go monolith reusing osv.dev microservices via goroutine-per-service architecture.

    ### Authentication
    - **Bearer JWT**: Obtained via `POST /api/v1/auth/login`
    - **API Key**: Pass `X-API-Key` header with a pre-issued key

    ### Scan Types
    | Type | Description |
    |------|-------------|
    | `discovery` | Fast host discovery (nmap ping scan) |
    | `full` | Full port + service scan |
    | `web` | Web application scan via OWASP ZAP |
    | `agent` | Agent-reported package vulnerability scan |
  version: "1.0.0"
  contact:
    email: admin@openvulnscan.local
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://openvulnscan.example.com
    description: Production

# ── Security Schemes ──────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Pre-issued API key for programmatic access
  schemas:
...
    AgentReportRequest:
      type: object
      required: [hostname, packages]
      properties:
        hostname:
          type: string
          example: "prod-server-01"
        os_info:
          type: string
          example: "Ubuntu 22.04"
        packages:
          type: array
          items:
            type: object
            required: [name, version]
            properties:
              name:
                type: string
              version:
                type: string
              ecosystem:
                type: string
                example: "PyPI"

    # ── Report ────────────────────────────────────────────────────────────────
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        scan_id:
          type: string
          format: uuid
        format:
          type: string
          enum: [json, csv, html, pdf]
        status:
          type: string
          enum: [pending, completed, failed]
        download_url:
          type: string
        generated_at:
          type: string
          format: date-time

    # ── SIEM ──────────────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/scans/{id}/report:
    get:
      tags: [Scans, Reports]
      summary: Get scan report (JSON)
      operationId: getScanReport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scan report data
    post:
      tags: [Scans, Reports]
      summary: Generate and download scan report
      operationId: generateScanReport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [json, csv, html]
                  default: json
      responses:
        "200":
          description: Report file
          content:
            application/json:
              schema:
                type: string
            text/csv:
              schema:
                type: string
            text/html:
              schema:
                type: string
...
  /api/v1/cves/{id}:
    get:
      tags: [Vulnerabilities]
      summary: Get CVE details from database
      operationId: getCVE
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: "CVE-2021-44228"
      responses:
        "200":
          description: CVE detail with CVSS scores, references, and affected packages
        "404":
          description: CVE not found
  /api/v1/cves:
    get:
      tags: [Vulnerabilities]
      summary: Search CVEs
      operationId: searchCVEs
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query (CVE ID, keyword)
        - in: query
          name: severity
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW]
        - in: query
          name: min_score
          schema:
            type: number
      responses:
        "200":
          description: CVE search results

  # ── Products ────────────────────────────────────────────────────────────────
...
  /api/v1/reports:
    get:
      tags: [Reports]
      summary: List generated reports
      operationId: listReports
      responses:
        "200":
          description: Report list
    post:
      tags: [Reports]
      summary: Generate a report for a scan
      operationId: generateReport
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [scan_id]
              properties:
                scan_id:
                  type: string
                  format: uuid
                format:
                  type: string
                  enum: [json, csv, html]
                  default: json
      responses:
        "201":
          description: Report generation queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
  /api/v1/reports/{id}:
    get:
      tags: [Reports]
      summary: Get report metadata
      operationId: getReport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
...
  /agent/report:
    post:
      tags: [Agent]
      security: []
      summary: Submit agent package inventory report
      description: |
        Agents submit package lists which are enriched via OSV API.
        Findings are created automatically for vulnerable packages.
      operationId: submitAgentReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentReportRequest'
            example:
              hostname: prod-server-01
              os_info: "Ubuntu 22.04.3 LTS"
              packages:
                - name: log4j
                  version: "2.14.0"
                  ecosystem: Maven
                - name: openssl
                  version: "1.1.1t"
                  ecosystem: Debian
      responses:
        "200":
          description: Report accepted for enrichment
...