# archive/cve-search-service/internal/delivery/http/handler.go
// Package http provides the HTTP delivery layer for the CVE search service.
package http

import (
	"encoding/json"
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/rs/zerolog"

	domainerrors "github.com/globalcve/cve-search-service/internal/domain/errors"
	"github.com/globalcve/cve-search-service/internal/domain/repository"
	"github.com/globalcve/cve-search-service/internal/usecase/getbyid"
	"github.com/globalcve/cve-search-service/internal/usecase/search"
)

// Handler holds HTTP handlers for the CVE search service.
type Handler struct {
	searchUC  *search.UseCase
	getByIDUC *getbyid.UseCase
	cveRepo   repository.CVERepository
	log       zerolog.Logger
}

// NewHandler creates an HTTP Handler.
func NewHandler(searchUC *search.UseCase, getByIDUC *getbyid.UseCase, cveRepo repository.CVERepository, log zerolog.Logger) *Handler {
	return &Handler{searchUC: searchUC, getByIDUC: getByIDUC, cveRepo: cveRepo, log: log}
}

// Health handles GET /health.
func (h *Handler) Health(w http.ResponseWriter, r *http.Request) {
	respondJSON(w, http.StatusOK, map[string]string{"status": "ok", "service": "cve-search-service"})
}

// SearchCVEs handles GET /api/v2/cves.
func (h *Handler) SearchCVEs(w http.ResponseWriter, r *http.Request) {
	req := &search.Request{
		Query:    r.URL.Query().Get("query"),
		Severity: strings.ToUpper(r.URL.Query().Get("severity")),
		Source:   strings.ToUpper(r.URL.Query().Get("source")),
		Sort:     r.URL.Query().Get("sort"),
		Page:     parseInt(r.URL.Query().Get("page"), 0),
		Limit:    parseInt(r.URL.Query().Get("limit"), 50),
	}

	resp, err := h.searchUC.Execute(r.Context(), req)
	if err != nil {
		h.log.Error().Err(err).Msg("search failed")
		respondError(w, http.StatusInternalServerError, "search failed")
		return
	}
	respondJSON(w, http.StatusOK, resp)
}

// GetCVE handles GET /api/v2/cves/{id}.
func (h *Handler) GetCVE(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	cve, err := h.getByIDUC.Execute(r.Context(), id)
	switch err {
	case nil:
		respondJSON(w, http.StatusOK, cve)
	case domainerrors.ErrInvalidCVEID:
		respondError(w, http.StatusBadRequest, "invalid CVE ID format")
	case domainerrors.ErrCVENotFound:
		respondError(w, http.StatusNotFound, "CVE not found")
	default:
		h.log.Error().Err(err).Str("id", id).Msg("get cve failed")
		respondError(w, http.StatusInternalServerError, "internal error")
	}
}

// GetCount handles GET /internal/cves/count.
func (h *Handler) GetCount(w http.ResponseWriter, r *http.Request) {
	n, err := h.cveRepo.Count(r.Context())
	if err != nil {
		respondError(w, http.StatusInternalServerError, "count failed")
		return
	}
	respondJSON(w, http.StatusOK, map[string]int64{"count": n})
}

// NewRouter creates the chi router for the CVE search service.
func NewRouter(h *Handler, log zerolog.Logger) http.Handler {
	r := chi.NewRouter()
	r.Use(middleware.RealIP)
	r.Use(middleware.Recoverer)
	r.Use(middleware.RequestID)
	r.Use(middleware.Timeout(30 * time.Second))
	r.Use(requestLogger(log))

	r.Get("/health", h.Health)
	r.Get("/api/v2/cves", h.SearchCVEs)
	r.Get("/api/v2/cves/{id}", h.GetCVE)
	r.Get("/internal/cves/count", h.GetCount)

	return r
}

func requestLogger(log zerolog.Logger) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			start := time.Now()
			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
			next.ServeHTTP(ww, r)
			log.Info().
				Str("method", r.Method).
				Str("path", r.URL.Path).
				Int("status", ww.Status()).
				Dur("latency", time.Since(start)).
				Msg("request")
		})
	}
}

func respondJSON(w http.ResponseWriter, code int, v interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	json.NewEncoder(w).Encode(v) //nolint:errcheck
}

func respondError(w http.ResponseWriter, code int, msg string) {
	respondJSON(w, code, map[string]string{"error": msg})
}

func parseInt(s string, def int) int {
	if n, err := strconv.Atoi(s); err == nil {
		return n
	}
	return def
}